refactor: move configuration building into own file config.go
